Flash messages

+++ src/Controller/UsersController.php
@@ -24,20 +24,22 @@ class UsersController extends AbstractController
     	$avatar = $user->getAvatar();
 
     	if($request->request->count() > 0) {
-    		if ($avatar) {
-				$uploader->removeFile($avatar, "userAvatar"); 
-    		}
         	$firstName = $request->request->get("firstName");
         	$lastName = $request->request->get("lastName");
         	$file = $request->files->get("avatarFile");
-        	$filename = $uploader->upload($file, "userAvatar");
-        	$user->setAvatar($filename)
+            if ($file) {
+                if ($avatar) {
+                    $uploader->removeFile($avatar, "userAvatar"); 
+                }
+                $avatar = $uploader->upload($file, "userAvatar");
+            }
+        	$user->setAvatar($avatar)
         		 ->setFirstName($firstName)
         		 ->setLastName($lastName);
 
             $manager->persist($user);
             $manager->flush();
-
+            $this->addFlash('success', 'Vos modifiations ont bien été pises en compte,<br><strong>Merci</strong>.');
             return $this->redirectToRoute("account");
         }
         return $this->render('users/account.twig', [
